refactor(lockfile,cmd): give each half of the lock its own file
The lock had grown into two files doing four jobs each, which is fine until
somebody has to change one of them. internal/lockfile now separates the parts
that answer different questions: path.go is which file a lock is for and which
file a name reaches, owner.go is the record naming a holder and the message a
waiter reads, registry.go is the process-wide bookkeeping that keeps a held lock
reachable and serializes two acquisitions inside one process, and lockfile.go is
left with acquisition itself.

cmd is split along the same seam: statelock.go keeps the policy - which
subcommands lock at all - and the wrapper that holds a lock while a command
runs; locktargets.go answers what each mutating command changes; configpath.go
resolves where a command's gup.json is, which is a question of its own because
the answer has to be settled once and reused rather than asked again.

Nothing moved between packages and no exported name changed, so this is
declarations relocated and file-level documentation added. The behavior is
covered by the tests that were already there and the ones added with the fixes
above; none of them needed a change to keep passing.
